% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_plot.R
\name{Cluster_plot}
\alias{Cluster_plot}
\title{plot on different resolution}
\arguments{
\item{work_object_Clst}{object with cluster}

\item{quiery_res}{resolution to be plotted}

\item{label}{whether to add labels}
}
\description{
plot on different resolution
}
