% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{To convert among gene nomenclature among ENSG, symbol, entrezid .etc}
\arguments{
\item{ENSG_list}{input list}

\item{from_type}{type of input list}

\item{type}{type of output list}

\item{return_list}{whether to return a list of gene (or a data frame insteat)}
}
\description{
To convert among gene nomenclature among ENSG, symbol, entrezid .etc
}
\details{
return either a list of input genes or a corresponding data frame
}
