% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creat_object.R.R
\name{Create_object}
\alias{Create_object}
\title{create a seurat object from count matrix}
\arguments{
\item{project_name}{name of project}

\item{filter}{one of auto, manual, no(NO), 0; auto}

\item{...}{other parameter of CreatSuratObject, min.cell, min.features}
}
\description{
create a seurat object from count matrix
}
