% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{daily}
\alias{daily}
\title{generate daily file fastly and easily}
\arguments{
\item{date}{date}

\item{name}{name or title of today's work}

\item{project}{which project is the work is under, currently available (Module_correlation, Remove_batch, Dist_algorithm, data_generation)}
}
\description{
generate daily file fastly and easily
}
\details{
just generate 2 files (data and scripe)
}
