% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_generation.R.R
\name{DEG_generation}
\alias{DEG_generation}
\title{generate differential expressed gene}
\arguments{
\item{optmz_res}{resolution chosen for DEG genetion}

\item{p_thr}{p value threshold}

\item{logFC_thr}{logFC threshold}

\item{test_use}{default "wilcox" details from FindMarkers}

\item{only.pos}{only return the positions}
}
\description{
generate differential expressed gene
}
