% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim_red_Cluster.R
\name{Dim_red_Cluster}
\alias{Dim_red_Cluster}
\title{Dim reduction and find cluster of Seurat object}
\arguments{
\item{work_object}{Seurat object}

\item{npcs}{default as 20}

\item{...}{other parameter of FindVariableFeatures}
}
\description{
Dim reduction and find cluster of Seurat object
}
\details{
to perform dim reduction and cluster manually
}
