% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_AUC.R
\name{Draw_AUC}
\alias{Draw_AUC}
\title{plot on object generated by pROC::roc}
\arguments{
\item{rocobj}{an object generated by pROC::roc}

\item{title}{title of the plot}
}
\description{
plot on object generated by pROC::roc
}
\details{
plot on AUC
}
