% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Four_fold_change.R
\name{Four_fold_change}
\alias{Four_fold_change}
\title{add Sensitivity, NPV, accuracy, etc. following the four-fold table}
\arguments{
\item{Four_fold}{a dataframe, colnames as(c("Predict", "True")), Predict has c("Negative", "Positive"), True has c("Ctrl", "Disease")}
}
\description{
add Sensitivity, NPV, accuracy, etc. following the four-fold table
}
\details{
add Sensitivity, NPV, accuracy, etc. following the four-fold table
}
