% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Four_fold_rocobj.R
\name{Four_fold_rocobj}
\alias{Four_fold_rocobj}
\title{for a contructed rocobj, add Sensitivity, NPV, accuracy, etc. following the four-fold table}
\arguments{
\item{rocobj}{a rocobj generated by a dataframe, colnames as(c("Predict", "True")), Predict has c("Negative", "Positive"), True has c("Ctrl", "Disease")}
}
\description{
for a contructed rocobj, add Sensitivity, NPV, accuracy, etc. following the four-fold table
}
\details{
for a contructed rocobj, add Sensitivity, NPV, accuracy, etc. following the four-fold table
}
