% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_plot.R
\name{GO_plot}
\alias{GO_plot}
\title{return a GO plot or GO object}
\arguments{
\item{Gene}{gene to plot}

\item{GO_result}{a dataframe of GO result}

\item{GO}{a returned ego from return_ego = T}

\item{input_type}{type of gene input}

\item{showCategory}{number of categories to plotted}

\item{ignore_gene}{defult as F, wether to ignore some GO item related to the character described in ignore_list}

\item{ignore_list}{default as "RNA", "catabolic", "metabolic", "viral", "translation initiation", "protein target"}

\item{title}{title to add on GO plot}

\item{label_format}{default as 30}

\item{font.size}{default as 12}

\item{return_ego}{default as F}
}
\description{
return a GO plot or GO object
}
\details{
return a GO plot or GO object
}
