% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\title{glance at the data}
\usage{
header(df, x = 8, y = 8)
}
\arguments{
\item{df}{the 2d-matrix or dataframe}

\item{x}{dimension 1 to be returned}

\item{y}{dimension 2 to be returned}
}
\description{
glance at the data
}
\details{
return 8X8 if dimension < 8X10
}
