% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_list.R
\name{mutate_list}
\alias{mutate_list}
\title{convert list x into y through the map relationship x_cond --> y_cond}
\arguments{
\item{xlist}{xlist}

\item{x_cond}{x_cond}

\item{y_cond}{y_cond}

\item{force_1st}{if there are some replication in x_cond, whether to use the 1st element occured in x_cond to match}

\item{use_origin}{if one element in xlist is not present in x_cond, then use the origin element in xlist(T) or use NA(F), default as F}
}
\description{
convert list x into y through the map relationship x_cond --> y_cond
}
\details{
return a list as long as the length of x_cond = y_cond, if a element is present in x_list but absent in x_cond, return NA in y_list
}
