% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plot.R
\name{QC_plot}
\alias{QC_plot}
\title{plot on QC quality}
\arguments{
\item{input}{can be count matrix or Seurat object}

\item{project_name}{specify a project_name if input a count matrix}

\item{group}{use one group to split the qc plot}

\item{organism}{default to "hs"; can also be set to "ms", so that also mitochondrial genes are filtered as "^mt-"}
}
\description{
plot on QC quality
}
\details{
plot on QC quality
}
