% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_subset.R
\name{QC_subset}
\alias{QC_subset}
\title{return a subset of Surat object}
\arguments{
\item{percent.mt_high}{default = 100, no filter on percent of mt gene}

\item{Assign}{whether to assign the name of cell or gene, if T, then cell_assign_target and gene_assign_target should be assigned}
}
\description{
return a subset of Surat object
}
\details{
return 8X8 if dimension < 8X10
}
