% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchPM.R
\name{SearchPM}
\alias{SearchPM}
\title{search PM result}
\arguments{
\item{keyword}{search the keyword}

\item{start_pages}{num of page start to search}

\item{num_pages}{end of page to search}

\item{years}{a contiuous year, like c(2022:2024)}

\item{dir_save}{dir to save}

\item{returnres}{return the result apart from saving the results}

\item{returnURL}{return the URL to check the page nums}
}
\description{
search PM result
}
\details{
return a data.frame, better to check the page numbers (using return URL = T) before searching
}
