% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_URL.R
\name{try_URL}
\alias{try_URL}
\title{try to open a URL link}
\arguments{
\item{URL}{the link}

\item{upperlimit}{max time to try}

\item{content}{expressed on the screen}
}
\description{
try to open a URL link
}
\details{
return read_html(URL)
}
